setup multi routes
