+++ frontend/package.json
@@ -8,10 +8,10 @@
   "type": "module",
   "dependencies": {
     "@ant-design/icons": "^5.2.6",
-    "@ant-design/pro-layout": "^7.17.16",
+    "@ant-design/pro-layout": "^7.17.19",
     "@reduxjs/toolkit": "^2.0.1",
     "@vitejs/plugin-react": "^4.2.1",
-    "antd": "^5.12.8",
+    "antd": "^5.13.2",
     "antd-phone-input": "^0.3.3",
     "axios": "^1.6.2",
     "cross-env": "7.0.3",
@@ -22,11 +22,11 @@
     "react-dom": "^18.2.0",
     "react-quill": "^2.0.0",
     "react-redux": "^9.0.4",
-    "react-router-dom": "^6.21.1",
+    "react-router-dom": "^6.21.3",
     "redux": "^5.0.1",
     "reselect": "^5.1.0",
     "shortid": "^2.2.16",
-    "vite": "^5.0.11"
+    "vite": "^5.0.12"
   },
   "scripts": {
     "dev": "vite",
